+++ src/Entity/Core/Route.php
@@ -37,7 +37,7 @@ use Symfony\Component\Validator\Constraints as Assert;
  *         "generate"={ "method"="POST", "path"="/routes/generate" }
  *     },
  *     itemOperations={
- *         "get"={ "requirements"={"id"="(.+)"}, "security"="is_granted(object)" }
+ *         "get"={ "requirements"={"id"="(.+)"}, "security"="is_granted('read_route', object)" }
  *     }
  * )
  * @Assert\Expression(