+++ src/Security/EventListener/LogoutListener.php
@@ -13,7 +13,9 @@ declare(strict_types=1);
 
 namespace Silverback\ApiComponentsBundle\Security\EventListener;
 
+use Lexik\Bundle\JWTAuthenticationBundle\Security\Http\Cookie\JWTCookieProvider;
 use Silverback\ApiComponentsBundle\RefreshToken\Storage\RefreshTokenStorageInterface;
+use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Security\Http\Event\LogoutEvent;
 
 /**
@@ -22,14 +24,20 @@ use Symfony\Component\Security\Http\Event\LogoutEvent;
 class LogoutListener
 {
     private RefreshTokenStorageInterface $storage;
+    private JWTCookieProvider $cookieProvider;
 
-    public function __construct(RefreshTokenStorageInterface $storage)
+    public function __construct(RefreshTokenStorageInterface $storage, JWTCookieProvider $cookieProvider)
     {
         $this->storage = $storage;
+        $this->cookieProvider = $cookieProvider;
     }
 
     public function __invoke(LogoutEvent $event): void
     {
         $this->storage->expireAll($event->getToken()->getUser());
+        $response = $event->getResponse() ?? new Response();
+        $response->headers->setCookie($this->cookieProvider->createCookie('.', null, time()));
+        $response->headers->remove('Location');
+        $response->setStatusCode(Response::HTTP_OK)->setContent('');
     }
 }